pyproject.toml
[project]
name = "multi-agents-rag-cykg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "ipython>=9.3.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-neo4j>=0.4.0",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "mcp-use>=1.3.3",
    "numpy>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "rdflib>=7.1.4",
]
